Subscriber callbacks can have arbitrary parameters now.
Publisher::notify now supports passing a function pointer of a
subscriber class and it's parameter declaration. This results in
receiving compiler errors when attempting to pass an incorrect number of
parameters or incorrect type for the given function.

This is a pretty nice achievement if there is not any nuances
overlooked.