Added turnstile example.
The turnstile examples comes from the wiki page:
https://en.wikipedia.org/wiki/Finite-state_machine. Though that it
works, the states themselves do not operate on the object since they
hold no referenece to the object. This is problematic since the states
should control the behavior of the object. Next step is to figure out a
clean way to have the state perform operations on the object.