Initial commit of project: Timelog
This project is intended to show a functional program using the state
design API as its framework. So far, a timelog object has been defined
along with semantically meaningful alias. The implementation of the
functions are intentionally tersed in handling errors to ensure
correctness rather than worry about performance.

Currently, the timeline journal object is being worked on. What's being
considered is the fact that C++26 is needed for built-in hashing on time
objects, such as std::chrono::year_month_day. Though it is not needed,
it was intended to use for clarity semantically. Considerations are
being made on how due to needing to provide a custom hash for an
associative container, such as std::unordered_map.